Bug 1201669 - ensuring that logging is propagated from firefox host to its set destination.
