+++ test/launch_test.js
@@ -3,8 +3,10 @@ var serverHelper = require('./server_helper');
 
 marionette('basic', function() {
   var client = marionette.client({
-    prefs: {
-      'browser.shell.checkDefaultBrowser': false
+    profile: {
+      prefs: {
+        'browser.shell.checkDefaultBrowser': false
+      }
     }
   });
